requirements.txt
spotipy==2.23.0
ytmusicapi==1.6.0
python-dotenv==1.0.1
requests==2.32.3
numpy==1.26.4
scikit-learn==1.5.1
pandas==2.2.2
